Linux DRM support.
This provides wl_drm support to XWayland given a DRM device and
linux-dmabuf protocol support.

A DRM device needs to be provided using either --drm-device flag
or the XWL_DRM_DEVICE variable.

The default behavior is for XWayland to use software rendering
to DRM buffers when a DRM device is provided. The --glamor flag
can be used to have XWayland render using OpenGL instead.