Send all keys to client.
Use keyboard extension protocol and ack all key events as
handled. The result is that all keyboard shortcuts that are
not reserved will be sent to the application.

A list of exceptions will be introduced by a follow up change,
in order to allow certain keyboard shortcuts to still be
handled by the compositor.

Change-Id: I42c2b2533c9a1c7cb7bc844e15cd4feeb47ab2e1
Reviewed-on: https://chromium-review.googlesource.com/978646
Reviewed-by: David Reveman <reveman@chromium.org>
Tested-by: David Reveman <reveman@chromium.org>