Add reserved accelerators option.
This adds the --accelerators=ACCELERATORS flag and
SOMMELIER_ACCELERATORS setting. They makes it possible to specify
accelerators that should NOT be forwarded to clients.

The format of ACCELERATORS is a "," separated list of accelerator
descriptions. The description format is "|MODIFIERS|SYMBOL", where
MODIFIERS is an optional set of modifiers (a combination of
<Control> <Alt> and <Shift>) and SYMBOL is a XKB keysym name,
(XK_*) as defined in keysymdef.h.

Change-Id: Ie3ed86bcd9fbc6a9132963b3991f970212d52dca
Reviewed-on: https://chromium-review.googlesource.com/979332
Reviewed-by: David Reveman <reveman@chromium.org>
Tested-by: David Reveman <reveman@chromium.org>